// UaRestGateway/uarestgateway.client/src/opcua/models/SubscriptionAcknowledgement.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SubscriptionAcknowledgement
 */
export interface SubscriptionAcknowledgement {
    /**
     * 
     * @type {number}
     * @memberof SubscriptionAcknowledgement
     */
    SubscriptionId?: number;
    /**
     * 
     * @type {number}
     * @memberof SubscriptionAcknowledgement
     */
    SequenceNumber?: number;
}

/**
 * Check if a given object implements the SubscriptionAcknowledgement interface.
 */
export function instanceOfSubscriptionAcknowledgement(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SubscriptionAcknowledgementFromJSON(json: any): SubscriptionAcknowledgement {
    return SubscriptionAcknowledgementFromJSONTyped(json, false);
}

export function SubscriptionAcknowledgementFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubscriptionAcknowledgement {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'SubscriptionId': !exists(json, 'SubscriptionId') ? undefined : json['SubscriptionId'],
        'SequenceNumber': !exists(json, 'SequenceNumber') ? undefined : json['SequenceNumber'],
    };
}

export function SubscriptionAcknowledgementToJSON(value?: SubscriptionAcknowledgement | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'SubscriptionId': value.SubscriptionId,
        'SequenceNumber': value.SequenceNumber,
    };
}