// UaRestGateway/uarestgateway.client/src/opcua/models/DeleteSubscriptionsRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RequestHeader } from './RequestHeader';
import {
    RequestHeaderFromJSON,
    RequestHeaderFromJSONTyped,
    RequestHeaderToJSON,
} from './RequestHeader';

/**
 * 
 * @export
 * @interface DeleteSubscriptionsRequest
 */
export interface DeleteSubscriptionsRequest {
    /**
     * 
     * @type {RequestHeader}
     * @memberof DeleteSubscriptionsRequest
     */
    RequestHeader?: RequestHeader;
    /**
     * 
     * @type {Array<number>}
     * @memberof DeleteSubscriptionsRequest
     */
    SubscriptionIds?: Array<number>;
}

/**
 * Check if a given object implements the DeleteSubscriptionsRequest interface.
 */
export function instanceOfDeleteSubscriptionsRequest(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function DeleteSubscriptionsRequestFromJSON(json: any): DeleteSubscriptionsRequest {
    return DeleteSubscriptionsRequestFromJSONTyped(json, false);
}

export function DeleteSubscriptionsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeleteSubscriptionsRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'RequestHeader': !exists(json, 'RequestHeader') ? undefined : RequestHeaderFromJSON(json['RequestHeader']),
        'SubscriptionIds': !exists(json, 'SubscriptionIds') ? undefined : json['SubscriptionIds'],
    };
}

export function DeleteSubscriptionsRequestToJSON(value?: DeleteSubscriptionsRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'RequestHeader': RequestHeaderToJSON(value.RequestHeader),
        'SubscriptionIds': value.SubscriptionIds,
    };
}