// UaRestGateway/uarestgateway.client/src/opcua/models/BrowsePath.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPC UA REST API
 * This API provides simple REST based access to an OPC UA server.
 *
 * The version of the OpenAPI document: 1.05.4
 * Contact: office@opcfoundation.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RelativePath } from './RelativePath';
import {
    RelativePathFromJSON,
    RelativePathFromJSONTyped,
    RelativePathToJSON,
} from './RelativePath';

/**
 * 
 * @export
 * @interface BrowsePath
 */
export interface BrowsePath {
    /**
     * 
     * @type {string}
     * @memberof BrowsePath
     */
    StartingNode?: string;
    /**
     * 
     * @type {RelativePath}
     * @memberof BrowsePath
     */
    RelativePath?: RelativePath;
}

/**
 * Check if a given object implements the BrowsePath interface.
 */
export function instanceOfBrowsePath(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function BrowsePathFromJSON(json: any): BrowsePath {
    return BrowsePathFromJSONTyped(json, false);
}

export function BrowsePathFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrowsePath {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'StartingNode': !exists(json, 'StartingNode') ? undefined : json['StartingNode'],
        'RelativePath': !exists(json, 'RelativePath') ? undefined : RelativePathFromJSON(json['RelativePath']),
    };
}

export function BrowsePathToJSON(value?: BrowsePath | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'StartingNode': value.StartingNode,
        'RelativePath': RelativePathToJSON(value.RelativePath),
    };
}